==> notebooks/1a_generate-linkpred-graph.ipynb <==
from os.path import join as j
import littleballoffur
import pandas as pd
import networkx as nx
import random

# Load data

with open("../data_dirs.txt", "r") as f: # your dataset folder should  be in a text file here
    DATA_DIRS = f.readlines()
    DATA_DIRS = [p.strip() for p in DATA_DIRS]

APS_CIT_PATH = j(DATA_DIRS[0], "derived", "journal_citations", "aps_all_2010_citations.csv")

cols = ["CITED_DOI", "CITING_DOI", "CITING_YEAR", "CITED_YEAR"]
dtypes = dict(zip(cols, ["str"]*2+["int"]*2))
cit = pd.read_csv(APS_CIT_PATH, usecols=cols, dtype=dtypes)

cit

len(cit[cit.CITING_YEAR == 2010])

cit_2010 = cit[cit.CITING_YEAR == 2010]

# Merge with abstract data

APS_ABS_PATH = j(DATA_DIRS[0], "raw", "WOS", "wos_aps_abstracts.bz2")

abstracts = pd.read_csv(APS_ABS_PATH, dtype="str", index_col="DOI")

abstracts

pd.isna(abstracts).sum()

abstracts = abstracts[(~pd.isna(abstracts.ABSTRACT)) & (abstracts.ABSTRACT != "")]

abstracts.index.nunique()

abstracts.reset_index().DOI.value_counts()

abstracts.loc["10.1103/PhysRevD.58.024017"]

len(abstracts.drop_duplicates())

len(abstracts.reset_index().drop_duplicates(subset="DOI"))

abstracts = abstracts.reset_index().drop_duplicates(subset="DOI", keep="first").set_index("DOI") # why the duplicates?

cit.merge(abstracts, left_on="CITING_DOI", right_index=True).merge(abstracts, left_on="CITED_DOI", right_index=True)

cit_abs = cit.merge(abstracts, left_on="CITING_DOI", right_index=True).merge(abstracts, left_on="CITED_DOI", right_index=True)

cit_abs = cit_abs.rename(columns={"ABSTRACT_x": "CITING_ABSTRACT", "ABSTRACT_y": "CITED_ABSTRACT"})

len(cit_abs)/len(cit)

cit_abs_2010 = cit_abs[cit_abs.CITING_YEAR == 2010]

len(cit_abs_2010)

# Generate graph

1. Create citation network G of APS papers published **up to** 2010
2. Sample negative edges: sample x% of negative citing edges from 2010 papers
3. Sample existing edges where the citing paper was publised _in_ 2010
5. Get largest connected component, and final list of added/removed edges

G = nx.DiGraph()
G.add_edges_from(cit_abs[["CITING_DOI", "CITED_DOI"]].values)

len(G.nodes())

len(cit_abs.CITING_DOI.append(cit_abs.CITED_DOI).unique())

len(G.edges())

len(cit_abs.drop_duplicates(subset=["CITING_DOI", "CITED_DOI"]))

REMOVE_EDGE_PCT = 0.5
REMOVE_EDGE_NO = int(len(cit_abs_2010) * REMOVE_EDGE_PCT)

REMOVE_EDGE_NO

## Remove true edges

to_remove = cit_abs_2010.sample(REMOVE_EDGE_NO)

len(to_remove)

### Check removed nodes/edges

cit_abs_2010

to_remove

### Remove edges from graph

Gr = G.copy()
Gr.remove_edges_from(to_remove[["CITING_DOI", "CITED_DOI"]].values)

len(G.edges()) - len(Gr.edges())

## Check largest CC

len(max(nx.weakly_connected_components(Gr), key=len))

len(G.nodes())

len(G.edges())

Gcc = Gr.subgraph(max(nx.weakly_connected_components(Gr), key=len)).copy()

len(Gcc.nodes())

len(Gcc.edges())

len(Gr.nodes())

len(Gr.edges())

**let's use the full graph with removed edges (not the GCC)**

# Final list of edges

removed_edges = list(set(G.edges()) - set(Gcc.edges())) # edges that are in the true network but not in the one we'll use for link prediction

len(removed_edges)

removed_edges = pd.DataFrame(data=removed_edges, columns=["CITING_DOI","CITED_DOI"])

to_remove.merge(removed_edges)

to_remove

to_remove = to_remove.drop(labels=["CITING_YEAR", "CITED_YEAR"], axis=1)

#for_embed = pd.DataFrame(list(Gcc.edges()), columns=["CITING_DOI","CITED_DOI"])
for_embed = pd.DataFrame(list(Gr.edges()), columns=["CITING_DOI","CITED_DOI"])

for_embed

for_embed.merge(cit_abs[["CITING_DOI","CITED_DOI", "CITING_ABSTRACT", "CITED_ABSTRACT"]], on=["CITING_DOI", "CITED_DOI"])

for_embed = for_embed.merge(cit_abs[["CITING_DOI","CITED_DOI", "CITING_ABSTRACT", "CITED_ABSTRACT"]], on=["CITING_DOI", "CITED_DOI"])

to_remove

for_embed.append(to_remove)

len(for_embed.append(to_remove).drop_duplicates())

# Save

NETWORK_PATH = j(DATA_DIRS[0], "derived", "journal_citations", "link_prediction", "abstract_aps_all_2010_link_pred_{}-pct_neg-0.gz".format(int(REMOVE_EDGE_PCT*100)))
REMOVE_PATH = j(DATA_DIRS[0], "derived", "journal_citations", "link_prediction", "abstract_aps_all_2010_link_pred_{}-pct_neg-0_pos.csv".format(int(REMOVE_EDGE_PCT*100)))

for_embed = for_embed.astype(str)
to_remove = to_remove.astype(str)

for_embed.to_csv(NETWORK_PATH, header=True, index=False)

to_remove.to_csv(REMOVE_PATH, header=True, index=False)

len(to_remove)
