==> notebooks/1b_embed-linkpred-text.ipynb <==
import numpy as np
from os.path import join as j
import pandas as pd
import string
import sys
import torch
import utils
from tqdm import tqdm
from transformers import AutoModel, AutoTokenizer, BertModel, BertTokenizer
from sentence_transformers import SentenceTransformer
from gensim.models import doc2vec

# Load data

with open("../data_dirs.txt", "r") as f: # your dataset folder should  be in a text file here
    DATA_DIRS = f.readlines()
    DATA_DIRS = [p.strip() for p in DATA_DIRS]

PCT = 50

LP_PATH = j(DATA_DIRS[0], "derived", "journal_citations", "link_prediction", "abstract_aps_all_2010_link_pred_{}-pct_neg-0.gz".format(PCT))
LP_REMOVE_PATH = j(DATA_DIRS[0], "derived", "journal_citations", "link_prediction", "abstract_aps_all_2010_link_pred_{}-pct_neg-0_pos.csv".format(PCT))

lp = pd.read_csv(LP_PATH, dtype="str")
lp_rm = pd.read_csv(LP_REMOVE_PATH, dtype="str")

len(lp) + len(lp_rm)

lp_all = lp.append(lp_rm)
lp_abs = lp_all[["CITING_DOI", "CITING_ABSTRACT"]].rename(columns={"CITING_DOI": "DOI", "CITING_ABSTRACT": "ABSTRACT"})
lp_abs = lp_abs.append(lp_all[["CITED_DOI", "CITED_ABSTRACT"]].rename(columns={"CITED_DOI": "DOI", "CITED_ABSTRACT": "ABSTRACT"}))

lp_abs

lp_abs = lp_abs.drop_duplicates()

lp_abs = lp_abs.set_index("DOI")

lp_abs

# SciBERT

# SciBERT embed
scibert_tokenizer = AutoTokenizer.from_pretrained(
    "allenai/scibert_scivocab_uncased", cache_dir="scibert_cache"
)
scibert_model = BertModel.from_pretrained(
    "allenai/scibert_scivocab_uncased",
    cache_dir="scibert_cache",
    output_hidden_states=True,
)
scibert_model = scibert_model.to("cuda:0")


def get_scibert_embedding(text):
    if isinstance(text, str):
        text = [text]
    # Encode the text, adding the (required!) special tokens, and converting to
    # PyTorch tensors.
    input_id = scibert_tokenizer(text, return_tensors="pt", add_special_tokens=True, padding=True)[
        "input_ids"
    ][:, :512]

    # Run the text through BERT, and collect all of the hidden states produced
    # from all 12 layers.
    with torch.no_grad():
        outputs = scibert_model(input_id.to("cuda:0"))
        hidden_states = outputs[2]  # Get the hidden states
        token_embeddings = torch.stack(
            hidden_states, dim=0
        )  # Concatenate all hidden layers into a big tensor
        token_embeddings = torch.squeeze(
            token_embeddings, dim=1
        )  # Remove the batch dimension
        if len(text) > 1:
            token_embeddings = token_embeddings.permute(1, 0, 2, 3)  # Permut
            last_layers = token_embeddings[:, -5:-1, :]
            embs = last_layers[:, :, 0, :].sum(axis = 1).cpu().numpy()
            #embs = last_layers.sum(axis=1).mean(axis=1).cpu().numpy()
        else:
            last_layers = token_embeddings[-5:-1, :]
            embs = last_layers[:, 0, :].sum(axis = 0).cpu().numpy().reshape((1, -1))
            #embs = last_layers.sum(axis=0).mean(axis=0).cpu().numpy().reshape((1, -1))
    return input_id.numpy(), embs

TEXT_COL = "ABSTRACT"
nchunks = np.ceil(lp_abs.shape[0] / 200)  # split data to fit into GPU memory
prog_bar1 = tqdm(total=nchunks)  # show progress

emb_chunks = []
doi_chunks = []
for chunk in np.array_split(lp_abs, nchunks):
    emb_chunks.append(get_scibert_embedding(chunk[TEXT_COL].values.tolist())[1])
    doi_chunks.append(chunk.index.values)
    prog_bar1.update()

emb = np.vstack(emb_chunks)
doi = np.concatenate(doi_chunks)

# Put into a DF
df = pd.DataFrame(index=doi, data=emb)

df

EMB_PROP = "abstract"
EMB_METHOD = "scibert"

EMB_PATH =  j(DATA_DIRS[0], "embedding", "link-prediction", "aps_all_2010_{}_{}_{}-pct_vector.gz".format(EMB_PROP, EMB_METHOD, PCT))

df.to_csv(EMB_PATH, index=True, header=True)

del df

# Sentence-BERT

model = SentenceTransformer('paraphrase-mpnet-base-v2', device='cuda:1')

embeddings = model.encode(lp_abs.ABSTRACT)

embeddings.shape

df = pd.DataFrame(data=embeddings, index=lp_abs.index)

df

EMB_METHOD = "sentencebert"

EMB_PATH =  j(DATA_DIRS[0], "embedding", "link-prediction", "aps_all_2010_{}_{}_{}-pct_vector.gz".format(EMB_PROP, EMB_METHOD, PCT))

df.to_csv(EMB_PATH, index=True, header=True)

del df, embeddings

# doc2vec

sentences = [doc2vec.TaggedDocument(doc, [i]) for i, doc in enumerate(lp_abs.ABSTRACT.values)]

model_d2v = doc2vec.Doc2Vec(sentences, vector_size=128, window_size=5, workers=8)

model_d2v.docvecs[len(lp_abs)-1]

vectors = []
for i in range(len(lp_abs)):
    vectors.append(model_d2v.docvecs[i])

vectors = np.vstack(vectors)

vectors.shape

d2v = pd.DataFrame(index=lp_abs.index, data=vectors, columns=range(128))

EMB_METHOD = "doc2vec"

EMB_PATH_2 = j(DATA_DIRS[0], "embedding", "link-prediction", "aps_all_2010_{}_{}_{}-pct_vector.gz".format(EMB_PROP, EMB_METHOD, PCT))

d2v.to_csv(EMB_PATH_2, index=True, header=True)
