==> notebooks/2a_generate-umap-all.ipynb <==
# UMAP of multiple embeddings

This notebook is for running UMAP on different embeddings of _the same_ set of papers

NB: in our previous iterations of UMAP, each embedding had a different set of sample papers

import os
import pandas as pd
import umap

# Load sample

SAMPLE_PATH = "data/derived/umap/sample_aps_all_2010_abstract_doc2vec_umap.csv" # replace with data folder

sample_df = pd.read_csv(SAMPLE_PATH).drop_duplicates(subset="DOI").set_index("DOI")
sample_df["PACS_CODE_1"] = sample_df["PACS_CODE_1"].astype(str)
sample_df = sample_df[["PACS_CODE_1"]] # this is all we really need
sample_df = sample_df.sample(n=25000)

len(sample_df)

# Load embeddings, run UMAP

EMB_SPEC = EMB_SPEC = ["undirected_leigenmap", "undirected_node2vec", "undirected_residual2vec", "title_doc2vec", "title_scibert", "title_sentencebert", "abstract_doc2vec", "abstract_scibert", "abstract_sentencebert"]
EMB_PATH = "data/embedding/vectors/aps_all_2010_{}_vector.gz" # replace with data folder
umap_dict = {}

for e in EMB_SPEC:
    emb = pd.read_csv(EMB_PATH.format(e), header=0, index_col="DOI").sort_index()
    DIM = emb.shape[1]
    emb.columns = list(range(0,DIM))
    emb = emb.merge(sample_df, left_index=True, right_index=True)
    
    reducer = umap.UMAP()
    umap_embedding = reducer.fit_transform(emb[list(range(0,DIM))].values)
    umap_df = pd.DataFrame(umap_embedding)
    umap_df.index = emb.index
    umap_df["PACS_CODE_1"] = emb["PACS_CODE_1"]
    umap_df = umap_df.rename(columns={0:"x", 1:"y"}).sort_index()
    umap_dict[e] = umap_df

umap_dict["title_scibert"].columns

umap_dict.keys()

[len(v) for v in umap_dict.values()]

umap_all = umap_dict["undirected_leigenmap"].merge(umap_dict["undirected_node2vec"][["x","y"]], left_index=True, right_index=True, suffixes=["_undirected_leigenmap", "_undirected_node2vec"])
umap_all = umap_all.merge(umap_dict["undirected_residual2vec"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_undirected_residual2vec", "y":"y_undirected_residual2vec"})
umap_all = umap_all.merge(umap_dict["title_doc2vec"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_title_doc2vec", "y":"y_title_doc2vec"})
umap_all = umap_all.merge(umap_dict["title_scibert"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_title_scibert", "y":"y_title_scibert"})
umap_all = umap_all.merge(umap_dict["title_sentencebert"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_title_sentencebert", "y":"y_title_sentencebert"})
umap_all = umap_all.merge(umap_dict["abstract_doc2vec"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_abstract_doc2vec", "y":"y_abstract_doc2vec"})
umap_all = umap_all.merge(umap_dict["abstract_scibert"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_abstract_scibert", "y":"y_abstract_scibert"})
umap_all = umap_all.merge(umap_dict["abstract_sentencebert"][["x","y"]], left_index=True, right_index=True).rename(columns={"x":"x_abstract_sentencebert", "y":"y_abstract_sentencebert"})

umap_all

# Save

umap_all.to_csv("sample_umap_all_embs.csv", index=True, header=True)
